backend/pyproject.toml
[project]
name = "eng_edu-backend"
version = "0.1.0"
description = "eng_edu API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "structlog>=24.1.0",
    "redis>=5.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
